Add ansible-junos-stdlib version 1.2.0.
This enables ansible to perform operations & config of JunOS boxes.

This package installs files in PKG_SYSCONFDIR, which pkgsrc and
pkglint takes a dim view at -- pkglint says "use CONF_FILES framework".
This is however what "ansible-galaxy install Juniper.junos" would do,
and this package simply replicates that behaviour for now.

So ... this is for review and/or improvement.