Import Lugaru from Wolfire Games as wip/lugaru.
The game compiles OK on NetBSD, but I think the patch provided here breaks
the installation procedure on OS X. Nevertheless, this is yet another one of
those programs where the game's data is expected to be in a path relative to
the main executable. wiz@ has suggested a workaround which I will apply to
this and many other games in the near future.