tigervnc: Import tigervnc-1.6.0 as wip/tigervnc
The package net/tigervnc (which is 1.5.0) is currently broken. This one
could replace it.

This package builds against graphics/jpeg instead of
graphics/libjpeg-turbo, which can be selected by setting
JPEG_DEFAULT=libjpeg-turbo.

This package has --disable-glx, because the required graphics/Mesa and
graphics/MesaLib packages are currently broken.

TigerVNC is a high-performance, platform-neutral implementation of VNC
(Virtual Network Computing), a client/server application that allows users
to launch and interact with graphical applications on remote machines.
TigerVNC provides the levels of performance necessary to run 3D and video
applications, and it attempts to maintain a common look and feel and
re-use components, where possible, across the various platforms that it
supports. TigerVNC also provides extensions for advanced authentication
methods and TLS encryption.