abduco: Import abduco-0.6 as wip/abduco
abduco provides session management i.e. it allows programs to be run
independently from their controlling terminal. That is programs can be detached
- run in the background - and then later reattached. Together with dvtm it
provides a simpler and cleaner alternative to tmux or screen.

abduco is in many ways very similar to dtach but is a completely independent
implementation which is actively maintained, contains no legacy code, provides
a few additional features, has a cleaner, more robust implementation and is
distributed under the ISC license.