kdewebkit: Import kdewebkit-5.21.0
This library provides KDE integration of the QtWebKit library. If you are
using QtWebKit in your KDE application, you are encouraged to use this layer
instead of using the QtWebKit classes directly.

In particular, you should use KWebView in place of QWebView, KGraphicsWebView
in place of QGraphicsWebView and KWebPage in place of QWebPage.  See the
documentation for those classes for more information.