kxmlrpcclient: Import kxmlrpcclient-4.21.0
This library contains simple XML-RPC Client support. It is a complete
client and is quite easy to use. Only one interface is exposed to the
world, kxmlrpcclient/client.h and of that interface, you only need to
use 3 methods: setUrl, setUserAgent and call.

A small note on authentication. If you will be accessing an XML-RPC server
which uses HTTP-AUTH, simply set the user and pass in the URL. To use
Digest authentication, call setDigestAuthEnabled(true).