Import parallel-20151122 as wip/parallel.
GNU Parallel is a shell tool for executing jobs in parallel using one or
more machines.  A job is typically a single command or a small script that
has to be run for each of the lines in the input.  The typical input is a
list of files, a list of hosts, a list of users, or a list of tables.