Import /home/bacon/Pkgsrc/pkgsrc-2016Q1/wip/slurm slurm-15.08.11 as wip/slurm.
Replaces old wip version 14.11.6.

SLURM is an open-source resource manager designed for Linux
clusters of all sizes. It provides three key functions. First it
allocates exclusive and/or non-exclusive access to resources
(computer nodes) to users for some duration of time so they can
perform work. Second, it provides a framework for starting,
executing, and monitoring work (typically a parallel job) on a
set of allocated nodes. Finally, it arbitrates contention for
resources by managing a queue of pending work.