Import gnome-core-3.20.2 as wip/gnome-core.
GNOME: The Free Software Desktop Project

GNOME 3 is an easy and elegant way to use your computer. It is
designed to put you in control and bring freedom to everybody.
GNOME 3 is developed by the GNOME community, a diverse, international
group of contributors that is supported by an independent, non-profit
foundation.

This package contains the standard GNOME libraries.