Import rofi-1.0.1 as wip/rofi.
A popup window switcher roughly based on superswitcher. This version started
off as a clone of simpleswitcher, the version from Sean Pringle. All credit for
this great tool should go to him. Rofi developed extra features, like a
run-dialog, ssh-launcher and can act as a drop-in dmenu replacement, making it
a very versatile tool.

Using Rofi is a lot like dmenu, but extended for an improved work flow.