add u-boot-full
This package installs the U-Boot platform-independent tools and builds
the U-Boot loader for the target board specified via the mk.conf
variable UBOOT_TARGET.

The tools are installed with the 'u-boot-' prefixed to their names,
i.e. tools/mkimage is installed as ${PREFIX}/bin/u-boot-mkimage.

It does not install the loader itself, as the files and procedure
involved can vary from target to target. Instead, while not ideal, it
leaves these files in the work tree for the user to install manually.

Additionally modified versions of NetBSD's stdarg.h, stdbool.h
and stddef.h are included to allow U-Boot to build on NetBSD hosts, as
well as a patch to U-Boot's default Kconfig to enable the option
CONFIG_USE_PRIVATE_LIBGCC for the same purpose (as well as reducing
reliance on any specific host toolchain).