datamash: add new package
Import it into pkgsrc if you find it useful.

GNU Datamash is a command-line program which performs basic
numeric,textual and statistical operations on input textual data
files.

it is designed to be portable and reliable, and aid researchers to
easily automate analysis pipelines, without writing code or even
short scripts.