Import barrage-1.0.4 as wip/barrage.
Barrage is a rather destructive action game that puts you on a shooting range
with the objective to hit as many dummy targets as possible within 3 minutes.
You control a gun that may either fire small or large grenades at soldiers,
jeeps and tanks. Targets are fast paced and require careful aiming as it takes
half a second to reload the gun. In addition, you have to keep an eye on your
ammo and reload in time.