letskencrypt: update to 0.1.9
2016-07-24: version 0.1.9

Have the system return an error code depending on what happened and
not just a blanket "success" (0) or "failure" (1). Now, the success
exit code (0) is only returned when certificates have been revoked
or updated. If no certificates were touched (e.g., they aren't out
of date and -F wasn't provided), then the system will return code
2. This allows the caller to know (for example) when the web server
should reload its certificates.

Have -N and -n only create the key if it does not exist. Allow
downstream to override PATH_VAR_EMPTY.

2016-07-12: version 0.1.8

Bug-fix for the incorrect validation of domain names as noted by
Remco and on GitHub — thanks! Add the -N flag to create the domain
key. This was indirectly requested in issues/7 and issues/6. Changed
to using SHA512 instead of SHA256 (it always was, but was being
incorrectly specified as SHA256) for the download file (noted in
issues/2). Fixed up letskencrypt(1) with some examples.