Import violetland-0.5 as wip/violetland.
Violetland is an open source cross-platform game similar to Crimsonland.

In this game the player should help a girl by name of Violet to struggle with
hordes of monsters. For this purpose the various weapon, and also the special
abilities of the heroine which are opening with experience can be used.