update to QCad 3.15.4
I started this package independently, then noticed that this already
existed in wip - I think I've got further along but still its not complete.
At least the application starts now and displays the splash screen.

I've got rid of Makefile.common - the other packages for QCad2 are not relevant
I think - the manual is online and the parts library is included in QCad3

the problem that I've got is detailed in the TODO, basically libtool is
refusing to install a few libraries for some reason.