qcad: fix more libtool issues
rework this. The libtool problems are because QCad build is set
up to link directly to a release/ directory. The only way I can
find to use the correct -rpath is to take DESTDIR away from the
qmake files, and add an install target.
We need to make a fake qtscript.pro file because the Qt binding
is handled in a subdir depending on the Qt version number.

add 'MAKE_JOBS_SAFE=no' as dependencies are not handled