py-autokey: add new package.
AutoKey is a desktop automation utility for Linux and X11. It allows
the automation of virtually any task by responding to typed
abbreviations and hotkeys. It offers a full-featured GUI that makes
it highly accessible for novices, as well as a scripting interface
offering the full flexibility and power of the Python language.