Import virtualbox-0.0.0nb20160805 as wip/virtualbox-svn.
VirtualBox is a family of powerful x86 virtualization products for
enterprise as well as home use. Not only is VirtualBox an extremely
feature rich, high performance product for enterprise customers, it
is also the only professional solution that is freely available as
Open Source Software under the terms of the GNU General Public License.

WWW: http://www.virtualbox.org/

Reimport this as an SVN repository. The git mirror is back in time too much.

For now there are no patches, they will be recreated on top of the current
master from VCS with improved support for NetBSD.