Update to 2.1.0 version.
Using a hash to identify from github.
It identifies itself as 2.1.0.
I needed this for a testing tool for DNS ECS that didn't work with 2.0.0.

There is no changes log, but from github:

Allow setting the RD bit
tresni committed on Mar 8

Don't die if a resolver timesout on option
tresni committed on Mar 8

Use the correct scope when exporting to wire format
rgacogne committed on Mar 8