Import chromium-browser-54.0.2820.1 as wip/chromium-browser.
Chromium is an open-source browser project that aims to build a safer,
faster and more stable way for Internet users to experience the web.

This repository is intended for work with the latest upstream tag and
prototype patches for upstreaming NetBSD support.

Right now it fetches the sources from a tag without submodules, it will be
addressed later.