Update Go to 1.7rc6.
The latest Go release, version 1.7, arrives six months after 1.6.
Most of its changes are in the implementation of the toolchain, runtime,
and libraries.
There is one minor change to the language specification.
As always, the release maintains the Go 1 promise of compatibility.
We expect almost all Go programs to continue to compile and run as
before.

The release adds a port to IBM LinuxOne; updates the x86-64 compiler
back end to generate more efficient code; includes the context package,
promoted from the x/net subrepository and now used in the standard
library; and adds support in the testing package for creating
hierarchies of tests and benchmarks.  The release also finalizes the
vendoring support started in Go 1.5, making it a standard feature.