Import mono-4.4.0.122 as wip/mono38.
The Mono Project is an open development initiative sponsored by Ximian
that is working to develop an open source, Unix version of the Microsoft
.NET development platform.  Its objective is to enable Unix developers to
build and deploy cross-platform .NET Applications.  The project will
implement various technologies developed by Microsoft that have now been
submitted to the ECMA for standardization.

This package is designed for mono-3.8 the latest stable version capable
to build native mcs.