surf2-git: import surf2-0.0 as wip/surf2-git
surf is a simple web browser based on WebKit/GTK+. It is able to display
websites and follow links. It supports the XEmbed protocol which makes it
possible to embed it in another application. Furthermore, one can point
surf to another URI by setting its XProperties.

This version is based on the WebKit2 API.