Import smallbasic-gui-0.12.6 as wip/smallbasic-gui.
SmallBASIC is a fast and easy to learn BASIC language interpreter ideal for
everyday calculations, scripts and prototypes. SmallBASIC includes
trigonometric, matrices and algebra functions, a built in IDE, a powerful
string library, system, sound, and graphic commands along with structured
programming syntax.

This package ships with a GUI version of SmallBASIC.