Import py27-pattyrn-1.2 as wip/py-pattyrn.
PyPattyrn is a python package aiming to make it easier and faster to implement
design patterns into your own projects.

Design patterns by nature cannot be directly translated into code as they are
just a description of how to solve a particular problem. However, many of the
common design patterns have some boilerplate code that is common throughout all
implementations of the pattern. This package captures that common code and
makes it easy to use so that you dont have to write it yourself in all your
projects.