bitcoin: Switch to upstream's install target
This commit removes large amounts of special-case code surrounding
secp256k1, and replaces it with running the install target in the
secp256k1 source directory.  It remains to understand why upstream's
main install target does not do this, should this issue survive
version upgrades.

Move bitcoind to bin from sbin, because that's where upstream puts it.