bitcoin: Drop dbus option
The dbus option added qt4-dbus, but upstream's current build
instructions do not mention this.  Without it, the package builds and
runs, so just drop it.

Add note that qt option should not be enabled by default, as it
doubles the size of the package.  (And, upstream appears to have some
split package support.)