lldb-netbsd: Document alternative test suits
Currently check-lldb is the only useful test target.

check-lldb-unit is not designed for a standalone mode
check-lldb-single = check-lldb -j1
check-lldb-expr does not seem to contain useful tests for now

check-lldb might be extended to be running explicitly in remote process setup.
Although this is not crucial as normal mode uses the remode approach internally
behind the scenes.

Sponsored by <The NetBSD Foundation>