First commit of clementine
As of this commit, building without options and building
with the suggested options should be fine. This was tested
(and built) on NetBSD on evbarm7hf. It wasn't actually tested
in X11 yet. I would like if someone were to try it and tell me
how that goes.

Errors are from the audiocd options and there were some more
that I couldn't reproduce. Those that were logged are in the
error/ directory. Since there was a less complete clementine
package on wip I'll be pushing it to clementine.old.

It's safe to assume that either the wiimote or gio options
cause an error, since building with all of the options except
pulseaudio, audiocd and lastfm (which doesn't build on ARM)
returned a different error twice (a linker error that I didn't log
and a compiler error that's in the errors dir).

The pulseaudio and lastfm options were never tested.