lldb-netbsd: Reintroduce NativeProcessNetBSD::SetDefaultPtraceOpts()
Add SetDefaultPtraceOpts() setting:
 - PTRACE_FORK
 - PTRACE_VFORK_DONE
 - PTRACE_LWP_CREATE
 - PTRACE_LWP_EXIT
with PT_SET_EVENT_MASK.

The PTRACE_VFORK event is currently disabled as unimplemented on NetBSD.

The purpose of this change is to be able to detect the above events in a
tracee in MonitorCallback.

Call SetDefaultPtraceOpts() in function launching inferior process and other.

While there, add more code to support (in future) attaching to a thread.
Currently it's not complete (PID 13555 is running top(1)):

$ lldb -p 13445
(lldb) process attach --pid 13445
error: attach failed: unable to attach
(lldb)

Functional attaching is planned to be added in next milestone.

Sponsored by <The NetBSD Foundation>