lldb-netbsd: Fix process resume action for NetBSD
Old report:

The current status is that in Linux process=thread and each
thread needs to be spawned or suspended separately. The code for remote
debugging is designed after Linux model and for NetBSD, we need to mimic
that there is single thread for certain interfaces (I've discussed it
with LLDB developers) - matching our concept of Process. Linux has code
to step or resume a process in the NativeThreadLinux part, we need to
call it per-process basis.

The action of Signal Monitor was ignored, as a tracee was marked as
Stopped after attaching (Launching -> Stopped). In the code to resume
it, I was just calling PT_CONTINUE without altering the status of tracee
(to Running or Stepping) and using ResumeAction list (it contains signal
to be passed). I discussed the proper design for NetBSD and our code for
it should live in NativeProcessNetBSD (not in NativeThreadNetBSD).

New report:

$ lldb
(lldb) process connect connect://localhost:1234
Process 29742 stopped
* thread #1, stop reason = The signal Stopped (signal) was caught
    frame #0:
(lldb) c
Process 29742 resuming
Hello world!
Process 29742 stopped
* thread #1, stop reason = The signal was generated via _lwp_kill(2) from pid=29742, uid=1000
    frame #0:
(lldb) c
Process 29742 resuming
Process 29742 exited with status = 0 (0x00000000)
(lldb)

Sponsored by <The NetBSD Foundation>