lldb-netbsd: Update TODO for next milestone; disable (v)fork monitor
Add a note on debugging the GDB remote protocol.

Remove entries for the "current milestone" as it's finished now.
Change "short-term goals in the next milestone" to the current ones.
Add two new ptrace(2) entries to the roadmap.

Disable monitor for (v)fork-like events, support for them is not planned for
the LLDB porting goal as of today.

Sponsored by <The NetBSD Foundation>