edbrowse: Update wip/edbrowse to edbrowse-3.6.2
pkgsrc changes:
 - instruct cmake to find tidy header files (in the code it directly include
   tidy header files without the <tidy/.*> path)
 - get rid of a local patch imported upstream

Changes:
3.6.2:
Read and convert utf16 and utf32 as per the byte order mark.
convert such files to utf8 or iso8859-1 as per the console setting,
though the latter is deprecated.
Convert back only if writing back out to the same file.

Enter high unicodes by the string ~u....;
where the dots are hex digits and the semi is optional.

New "list buffers" command (bfl)  to get a list of buffers and filenames.

Read the attachment filename from the http content-disposition header and set the current filename accordingly.
The url could be replaced with foobar.zip, but that's what the web designer wanted.

If a file has a recognized suffix, with a stream plugin,
then typing g on a link within that file invokes the same plugin.
This is primarily used for .pls files, which are playlists,
and each URL therein is intended for a music player.
Other browsers seem to work this way.

3.6.1:
Merge edbrowse and edbrowse-js back into one executable image.
More convenient for distribution.
The --mode argument determines the action of the process.
Example edbrowse --mode js to run as the js engine.

Simple implementation of xhr under javascript, synchronous only.

Don't encode urls that have already been encoded via <A href=url>.
The second encoding is mostly harmless, except for + becoming %2b.

Turn on or off the progress dots during an http or ftp download,
or receive progress counts by megabyte.

Create the edbrowse temp dir at the outset, and user directories
beneath this directory, mod 700, for added security in a multiuser system.
Temp files for plugins are created beneath the user directories.

Reload the config file on command. No need to exit and restart.

~0 in an edbrowse function is the whole line, even if more than 9 arguments.

db>filename to redirect debugging output to a file.