intel-vaapi-driver: Import intel-vaapi-driver-1.7.3 as wip/intel-vaapi-driver
VA-API (Video Acceleration API) user mode driver for Intel GEN
Graphics family.

VA-API is an open-source library and API specification, which
provides access to graphics hardware acceleration capabilities for
video processing. It consists of a main library and driver-specific
acceleration backends for each supported hardware vendor.

The current video driver backend provides a bridge to the GEN GPUs
through the packaging of buffers and commands to be sent to the
i915 driver for exercising both hardware and shader functionality
for video decode, encode, and processing.

XXX: Actually the package is marked as BROKEN because a proper way to install it
XXX: need to be implemented. The resulting i965_drv_video.so was manually tested
XXX: with mpv on NetBSD/amd64 7.99.53 and native X.org and seems working though.