PicoLisp: Import PicoLisp-17.2.24 as wip/PicoLisp
PicoLisp provides a 1-to-1 mapping of a clean and powerful
Lisp derivate, to a simple and efficient virtual machine.
It supports persistent objects as a first class data type,
resulting in a database system of Entity/Relation classes and a
Prolog-like query language tightly integrated into the system.

As an application server framework, PicoLisp provides for
   NoSQL Database Management:
      Index trees, Object local indexes, Entity/Relation
      classes, Pilog (PicoLisp Prolog) queries, Multi-user
      synchronization, DB Garbage collection, Journaling, Replication
   User Interface
      Browser GUI, (X)HTML/CSS, XMLHttpRequest/JavaScript
   Application Server
      Process management, Process family communication, XML I/O,
      Import/export, User administration, Internationalization,
      Security, Object linkage, Postscript/Printing