Import py27-dnspython-1.15.0 as wip/py-dnspython.
dnspython is a DNS toolkit for Python. It supports almost all record types. It
can be used for queries, zone transfers, and dynamic updates. It supports TSIG
authenticated messages and EDNS0.

dnspython provides both high and low level access to DNS. The high level
classes perform queries for data of a given name, type, and class, and return
an answer set. The low level classes allow direct manipulation of DNS zones,
messages, names, and records.

dnspython originated at Nominum where it was developed to facilitate the
testing of DNS software. Nominum has generously allowed it to be open
sourced under a BSD-style license.