libpthread-subts: update to 0.4.
pthread-stubs 0.4 is now available. This new release stops providing
stubs, but instead expects libc to do so. On platforms where this is
not the case, "-pthread" is injected via our pkg-config file instead.
This was done to make loading libpthread (indirectly) via dlopen()
work correctly in the presence of pthread-stubs.