coccinelle: Import coccinelle-1.0.6 as wip/coccinelle
Coccinelle is a program matching and transformation engine which provides
the language SmPL (Semantic Patch Language) for specifying desired matches
and transformations in C code. Coccinelle was initially targeted towards
performing collateral evolutions in Linux. Such evolutions comprise the
changes that are needed in client code in response to evolutions in library
APIs, and may include modifications such as renaming a function, adding a
function argument whose value is somehow context-dependent, and reorganizing
a data structure. Beyond collateral evolutions, Coccinelle is successfully
used (by us and others) for finding and fixing bugs in systems code.

XXX: This package is intended to be used to update devel/coccinelle in pkgsrc
XXX: but there are some issues still not properly investigated that need
XXX: to be solved in order to be able to update it. For more information please
XXX: give a look to coccinelle/TODO.

TODO: (about the TODO) try to describe the issues more in depth!